Refactor `ComputationTimeWithRefresh` into own file
